fix: move external-service-pill
