+++ apps/web/src/app/api/og/external-service/route.tsx
@@ -20,7 +20,7 @@ const INDEX_TITLE = "External Status";
 const INDEX_DESCRIPTION =
   "Check if your external providers are working properly";
 
-// Mirrors getPillStyle in @openstatus/ui external-service-pill so the OG label
+// Mirrors getPillStyle in (landing)/status/external-service-pill so the OG label
 // matches the on-page pill.
 function getStatus(args: { indicator: string; status: string }): {
   label: string;
